// ui/src/api/generated/model/loginKpiResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */

export interface LoginKpiResponse {
  total: number;
  successful: number;
  failed: number;
  unique_users: number;
}